add localizations section
